+++ app/src/main/java/com/swu/hyperventilationsyndrome/model/DataModel.kt
@@ -1,7 +1,7 @@
 package com.swu.hyperventilationsyndrome.model
 
 data class DataModel(
-    val an_p: Int,
-    val ap_p: Int,
-    val frequency: Int
+    val an_p: Double,
+    val ap_p: Double,
+    val frequency: Double
 )
